+++ src/offset_date_time.rs
@@ -417,9 +417,7 @@ impl OffsetDateTime {
         year
     }
 
-    /// Get the month of the date in the stored offset. If fetching both the
-    /// month and day, it is more efficient to use
-    /// [`OffsetDateTime::month_day`].
+    /// Get the month of the date in the stored offset.
     ///
     /// The returned value will always be in the range `1..=12`.
     ///
@@ -440,8 +438,7 @@ impl OffsetDateTime {
         self.date().month()
     }
 
-    /// Get the day of the date in the stored offset. If fetching both the month
-    /// and day, it is more efficient to use [`OffsetDateTime::month_day`].
+    /// Get the day of the date in the stored offset.
     ///
     /// The returned value will always be in the range `1..=31`.
     ///
@@ -515,16 +512,16 @@ impl OffsetDateTime {
     ///
     /// ```rust
     /// # use time_macros::datetime;
-    /// assert_eq!(datetime!("2019-01-01 0:00 UTC").week(), 1);
-    /// assert_eq!(datetime!("2020-01-01 0:00 UTC").week(), 1);
-    /// assert_eq!(datetime!("2020-12-31 0:00 UTC").week(), 53);
-    /// assert_eq!(datetime!("2021-01-01 0:00 UTC").week(), 53);
+    /// assert_eq!(datetime!("2019-01-01 0:00 UTC").iso_week(), 1);
+    /// assert_eq!(datetime!("2020-01-01 0:00 UTC").iso_week(), 1);
+    /// assert_eq!(datetime!("2020-12-31 0:00 UTC").iso_week(), 53);
+    /// assert_eq!(datetime!("2021-01-01 0:00 UTC").iso_week(), 53);
     /// ```
     ///
     /// This function is `const fn` when using rustc >= 1.46.
     #[const_fn("1.46")]
-    pub const fn week(self) -> u8 {
-        self.date().week()
+    pub const fn iso_week(self) -> u8 {
+        self.date().iso_week()
     }
 
     /// Get the weekday of the date in the stored offset.